requirements.txt
langchain
langchain-openai
langchain-community
langchain-huggingface
langchain-text-splitters
langchain-groq
langchain-chroma
pypdf
google-genai
google-adk
sentence-transformers
chromadb
pydantic
python-dotenv